Update color scheme in Vuetify
This commit moves theme.js to Vuetify's folder, as suggested in the
documentation[1].

It also rewrites the original theme.js file to follow Vuetify theme
schema (with an extra "color type" for text: 'txt') and applies it to
most of the current project.
File Buttons.vue was left with no modification as it is being replaced
by different component in 'src/components/buttons/'.

[1]: https://vuetifyjs.com/en/features/theme/#custom-theme-variants